Remove script
